+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/approvalThresholdResult/index.ts
@@ -1,0 +1,1 @@
+export { ApprovalThresholdResult, type IApprovalThresholdResultProps } from './approvalThresholdResult';