+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/index.ts
@@ -1,0 +1,13 @@
+import { ProposalDataListItemStructure as Structure } from './proposalDataListItemStructure';
+
+export const ProposalDataListItem = {
+    Structure,
+};
+
+export type {
+    IApprovalThresholdResult,
+    IMajorityVotingResult,
+    IProposalDataListItemStructureProps,
+    ProposalStatus,
+    ProposalType,
+} from './proposalDataListItemStructure';