+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/proposalDataListItemStructure/index.ts
@@ -1,0 +1,2 @@
+export { ProposalDataListItemStructure } from './proposalDataListItemStructure';
+export * from './proposalDataListItemStructure.api';